Phase D — dissolve flash: delete _flash.py, realize_flash folds into the assembler
`_flash.py` is gone. The warp-tier flash is now realized inside the generic assembler:
`_assemble.realize_flash` (placed beside `assemble_carry` as its fragment-tier specialization)
builds the streaming `CarryScope` and hands it to `assemble_carry` — the SAME tower path
matmul / reduce use. `assembly/010_assemble` dispatches to it purely off the `Schedule.carry`
representation; `warp_chain_eligible` moves into the `split/005_warp_chain` shim.

So flash now rides the standard pipeline end to end:
  split (shim) → seed_graph + Schedule.carry → 010_assemble → realize_flash → assemble_carry
with the carrier read off `block.carrier`, geometry off the logical `buffers`, the QK^T/P@V
cells lowered by `kernel/005`, and the softmax by `realize_fragment_softmax`.

Honest scope: the realization LOGIC is irreducibly flash-specific and was relocated (not
dissolved into matmul/reduce-shared machinery) — there is no fragment-tier online-softmax or
register-staged C→A handoff anywhere else, so no generic traversal produces it. What dissolved
is the *special-casing*: no `_flash.py` module, no `FlashSpec`, no `flash` marker, no separate
assembler entry — flash is dispatched off a generic `Schedule` field and built through the one
carry assembler. The C→A handoff (register fragment → smem → ldmatrix) is the lone
flash-specific authoring left, clearly isolated in `realize_flash`.

Green (GPU): full tests/compiler/ 1684 passed, 4 skipped (-O1); flash e2e reference/generated
at deployable -O3 (31). make lint clean.
